+++ include/mppt.h
@@ -12,7 +12,7 @@
 
 namespace MPPT {
 
-    void init(int baudRate);
+    void init(unsigned long baudRate);
 
     void read();
