+++ src/bmv.cpp
@@ -2,7 +2,7 @@
 
 namespace BMV {
 
-    void init(int baudRate) {
+    void init(unsigned long baudRate) {
         if (BMV_ENABLED)
             BmvSerial.begin(baudRate);
     }
